feat(tiny-editor): missing text patterns property (#87)
The text_patterns property is missing. I would want it to be so you can turn it off or change it in any way you want. I personally now have customized tiny editor for telegram markdown and this text_patterns messes it up a lot, I had to extend the TinyEditor class.